Run 3: First deployment that appears to operate correctly!!!!!!
After several hours on the wrong track, I traced the application
issue to the PackageLoader call. It worked locally because I
wasn't running WSGI so the application was automatically on the
system path when run. That didn't apply on Opalstack.

Several further wasyed hours later, having failed to make it
work in about five different ways, I switched to the more
malleable FileSystemLoader, and configured it to seek in the
directory that the wsgi.py module itself lives in, and configured
it to seek in the directory that the wsgi.py module itself lives in.

Once the uwsgi process was able to load the wsgi app it became
evident that it was trying to import an 'application' object,
I assigned the Flask app's `wsgi_app` attribute to the application
variable, and bingo!