Fix site installation and QR code generation.
Now appeaer to have a fundamentally sound site generation system.
It may or my not improve over their system, but I like the ability
to roll back to a previous release, and the move towards being
able to target any site. It's going to take a while to turn this
into a polished system.